Use Apksig library for signing K-Klock, this fixes translating K-Klock.
Also automatic code cleanup

+++ app/src/main/java/jadx/core/utils/android/ExtDataInput.java
@@ -85,7 +85,7 @@ public class ExtDataInput extends DataInputDelegate {
         int total = 0;
         int cur = 0;
 
-        while ((total < n) && ((cur = (int) super.skipBytes(n - total)) > 0)) {
+        while ((total < n) && ((cur = super.skipBytes(n - total)) > 0)) {
             total += cur;
         }
